Fix the BinaryFileTest on windows. The dummy file was trying to be deleted while the file handle was still open. Windows locks access to files while any handle is open which resulted in a file access error. Refs #2191

+++ Code/Mantid/Framework/Kernel/inc/MantidKernel/BinaryFile.h
@@ -118,7 +118,7 @@ public:
 
     // get the size of the file in bytes and reset the handle back to the beginning
     handle->seekg(0, std::ios::end);
-    size_t filesize = handle->tellg();
+    size_t filesize = static_cast<size_t>(handle->tellg());
     handle->seekg(0, std::ios::beg);
 
     // check the file is a compatible size
